Move mocha common utils to root level test/utils
