+++ framework/test/mocha/functional/http/get/accounts/multisignatures.js
@@ -16,11 +16,11 @@
 
 require('../../../functional');
 
-const Scenarios = require('../../../../common/scenarios');
+const Scenarios = require('../../../../../utils/multisig_scenarios');
 const accountFixtures = require('../../../../fixtures/accounts');
 const apiHelpers = require('../../../../../utils/api');
 const waitFor = require('../../../../../utils/wait_for');
-const SwaggerEndpoint = require('../../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../../utils/swagger_spec');
 
 const expectSwaggerParamError = apiHelpers.expectSwaggerParamError;
