+++ framework/test/mocha/functional/http/get/blocks.js
@@ -18,7 +18,7 @@ require('../../functional');
 
 const accountFixtures = require('../../../fixtures/accounts');
 const waitFor = require('../../../../utils/wait_for');
-const SwaggerEndpoint = require('../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../utils/swagger_spec');
 const apiHelpers = require('../../../../utils/api');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
