+++ framework/test/mocha/functional/http/get/node/transactions_pending.js
@@ -23,7 +23,7 @@ const {
 } = require('@liskhq/lisk-transactions');
 const apiHelpers = require('../../../../../utils/api');
 const randomUtil = require('../../../../../utils/random');
-const SwaggerEndpoint = require('../../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../../utils/swagger_spec');
 const accountFixtures = require('../../../../fixtures/accounts');
 const waitFor = require('../../../../../utils/wait_for');
 const {
