+++ framework/test/mocha/functional/http/get/node/transactions_ready.js
@@ -19,7 +19,7 @@ const Promise = require('bluebird');
 const { transfer } = require('@liskhq/lisk-transactions');
 const apiHelpers = require('../../../../../utils/api');
 const randomUtil = require('../../../../../utils/random');
-const SwaggerEndpoint = require('../../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../../utils/swagger_spec');
 const accountFixtures = require('../../../../fixtures/accounts');
 const {
 	getNetworkIdentifier,
