+++ framework/test/mocha/functional/http/post/2.delegate.js
@@ -21,7 +21,7 @@ const {
 	registerDelegate,
 	DelegateTransaction,
 } = require('@liskhq/lisk-transactions');
-const phases = require('../../../common/phases');
+const phases = require('../../../../utils/transaction_confirmation');
 const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../../utils/api');
 const randomUtil = require('../../../../utils/random');
