+++ framework/test/mocha/functional/http/post/3.votes.js
@@ -21,7 +21,7 @@ const {
 	registerDelegate,
 	castVotes,
 } = require('@liskhq/lisk-transactions');
-const phases = require('../../../common/phases');
+const phases = require('../../../../utils/transaction_confirmation');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
