+++ framework/test/mocha/functional/http/post/7.dapps.out_transfer.js
@@ -19,7 +19,7 @@ const BigNum = require('@liskhq/bignum');
 const { transfer, createDapp } = require('@liskhq/lisk-transactions');
 const Promise = require('bluebird');
 const accountFixtures = require('../../../fixtures/accounts');
-const phases = require('../../../common/phases');
+const phases = require('../../../../utils/transaction_confirmation');
 const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../../utils/wait_for');
 const elements = require('../../../../utils/elements');
