+++ framework/test/mocha/functional/http/put/node.js
@@ -16,7 +16,7 @@
 
 require('../../functional');
 const genesisDelegates = require('../../../data/genesis_delegates.json');
-const SwaggerEndpoint = require('../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../utils/swagger_spec');
 const apiHelpers = require('../../../../utils/api');
 
 const expectSwaggerParamError = apiHelpers.expectSwaggerParamError;
