+++ framework/test/mocha/functional/ws/transport/transport.js
@@ -17,9 +17,9 @@
 require('../../functional');
 const { P2P } = require('@liskhq/lisk-p2p');
 const { transfer } = require('@liskhq/lisk-transactions');
-const { generatePeerHeader } = require('../../../common/generatePeerHeader');
+const { generatePeerHeader } = require('../../../../utils/generatePeerHeader');
 const waitFor = require('../../../../utils/wait_for');
-const SwaggerEndpoint = require('../../../common/swagger_spec');
+const SwaggerEndpoint = require('../../../../utils/swagger_spec');
 const randomUtil = require('../../../../utils/random');
 const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../../utils/api');
