+++ framework/test/mocha/integration/app.js
@@ -16,7 +16,7 @@
 
 const BigNum = require('@liskhq/bignum');
 const { getAddressFromPublicKey } = require('@liskhq/lisk-cryptography');
-const application = require('../common/application');
+const application = require('../../utils/application');
 const QueriesHelper = require('../common/integration/sql/queries_helper');
 const accountsFixtures = require('../fixtures/accounts');
 const roundsFixtures = require('../fixtures/rounds').rounds;
