+++ framework/test/mocha/integration/blocks/process/on_receive_block.js
@@ -26,7 +26,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
 const genesisDelegates = require('../../../data/genesis_delegates.json')
 	.delegates;
-const application = require('../../../common/application');
+const application = require('../../../../utils/application');
 
 const { ACTIVE_DELEGATES, BLOCK_SLOT_WINDOW } = global.constants;
