+++ framework/test/mocha/integration/transactions/4_multisignature/4_4_multisignature.js
@@ -18,7 +18,7 @@ const {
 	registerMultisignature,
 	createSignatureObject,
 } = require('@liskhq/lisk-transactions');
-const Scenarios = require('../../../common/scenarios');
+const Scenarios = require('../../../../utils/multisig_scenarios');
 const localCommon = require('../../common');
 const {
 	getNetworkIdentifier,
